+++ CHANGELOG.md
@@ -48,6 +48,11 @@
   [Eric Horacek](https://github.com/erichoracek)
   [#2121](https://github.com/realm/SwiftLint/issues/2121)]
 
+* Add a new `multiline_function_chains` rule to validate chained function 
+  calls start either on the same line or one per line.  
+  [Eric Horacek](https://github.com/erichoracek)
+  [#2214](https://github.com/realm/SwiftLint/issues/2121)]
+
 * Improves the `mark` rule's autocorrection.  
   [Eric Horacek](https://github.com/erichoracek)